+++ frontend/src/routes/admin/+page.server.ts
@@ -1,7 +1,7 @@
 import type { PageServerLoad, Actions } from './$types.js';
 import { fail } from '@sveltejs/kit';
 import { superValidate } from 'sveltekit-superforms';
-import { addProductSchema } from './schema';
+import { addProductSchema } from '$lib/api/adminSchema';
 import { zod } from 'sveltekit-superforms/adapters';
 
 export const load: PageServerLoad = async () => {
